pyproject.toml
[tool.poetry]
name = "spkemb_ecapa2"
version = "0.1.0"
description = "Project of speaker verfication by ecapa2"
#license = "MIT"
authors = ["Kai Washizaki"]
maintainers = ["washizaki kai <bandad.kw@gmail.com>"]
readme = "README.md"
keywords = [
    "Python",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: GPU :: NVIDIA CUDA",
    #"License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: Japanese",
    "Programming Language :: Python",
    "Topic :: Software Development",
]

packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
audiomentations = "^0.30.0"
gdown = "^4.7.1"
hydra-core = "^1.3.1"
joblib = "^1.2.0"
librosa = "^0.9.2"
matplotlib = "^3.7.1"
pandas = "^2.0.1"
python-dotenv = "^1.0.0"
pytorch-lightning = "^2.0.2"
scikit-learn = "^1.2.1"
soundfile = "^0.11.0"
sox = "^1.4.1"
tensorboard = "^2.13.0"
timm = "^0.9.2"
#torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp311-cp311-linux_x86_64.whl"}
#torchaudio = {url = "https://download.pytorch.org/whl/cu118/torchaudio-2.0.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
#torchvision = {url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
torch = {url = "https://download.pytorch.org/whl/cu121/torch-2.3.1%2Bcu121-cp311-cp311-linux_x86_64.whl"}
torchaudio = {url = "https://download.pytorch.org/whl/cu121/torchaudio-2.3.1%2Bcu121-cp311-cp311-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu121/torchvision-0.18.1%2Bcu121-cp311-cp311-linux_x86_64.whl"}
torchmetrics = ">0.11.1"
tqdm = "^4.65.0"
transformers = "^4.29.2"
wasabi = "~0.10.1"
numpy = "==1.*"
torch-audiomentations = "^0.11.1"
wandb = "^0.17.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mdformat-gfm = "^0.3.5"
mdformat-frontmatter = "^2.0.1"
mdformat_footnote = "^0.1.1"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = 3.11
# following setting is same as pysen
# https://github.com/pfnet/pysen/blob/main/setup.cfg#L12
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"